refactor: decodeInvoice into utils/invoice
